events progress

+++ coding/Events/event.h
@@ -1,9 +1,13 @@
 #ifndef EVENT_H
 #define EVENT_H
+#include <random>
+#include <iostream>
+#include <cmath>
+#include <chrono>
 
 class event {
     protected:
-    double likelihood;
+    int likelihood;
     int duration;
     int eventAge;
